some error handling

+++ intranet/apps/eighth/models.py
@@ -1110,9 +1110,10 @@ class EighthScheduledActivity(AbstractBaseEighthModel):
                                                            previous_activity_name=previous_activity_name,
                                                            previous_activity_sponsors=previous_activity_sponsors, own_signup=(user == request.user))
                     if previous_activity.waitlist.all().exists():
-                        next_wait = EighthWaitlist.objects.get_next_waitlist(previous_activity)
-                        previous_activity.add_user(next_wait.user)
-                        next_wait.delete()
+                        if not previous_activity.is_full():
+                            next_wait = EighthWaitlist.objects.get_next_waitlist(previous_activity)
+                            previous_activity.add_user(next_wait.user)
+                            next_wait.delete()
 
                 except EighthSignup.DoesNotExist:
                     EighthSignup.objects.create_signup(user=user, scheduled_activity=self, after_deadline=after_deadline)
@@ -1336,9 +1337,10 @@ class EighthSignup(AbstractBaseEighthModel):
             block = self.scheduled_activity.block
             self.delete()
             if self.scheduled_activity.waitlist.all().exists():
-                next_wait = EighthWaitlist.objects.get_next_waitlist(self.scheduled_activity)
-                self.scheduled_activity.add_user(next_wait.user)
-                next_wait.delete()
+                if not self.scheduled_activity.is_full():
+                    next_wait = EighthWaitlist.objects.get_next_waitlist(self.scheduled_activity)
+                    self.scheduled_activity.add_user(next_wait.user)
+                    next_wait.delete()
             return "Successfully removed signup for {}.".format(block)
 
     def accept_pass(self):
